+++ pytket/tests/backendinfo_test.py
@@ -19,7 +19,7 @@ than a dataclass
 
 from json import dumps, loads
 
-from hypothesis import given
+from hypothesis import given, settings
 import pytest  # type: ignore
 
 from pytket.backends.backendinfo import BackendInfo, fully_connected_backendinfo
@@ -165,6 +165,7 @@ def test_to_json() -> None:
 
 
 @given(st.backendinfo())
+@settings(deadline=None)
 def test_backendinfo_serialization(backinfo: BackendInfo) -> None:
     serializable = backinfo.to_dict()
     assert BackendInfo.from_dict(serializable) == backinfo